requirements.txt
coverage==6.5.0
dvc==2.28.0
numpy==1.23.1
pandas==1.1.3
scikit-learn==1.1.1
pyyaml==5.3.1
joblib==1.3.2
fastapi==0.110.1
starlette==0.37.2
pydantic==2.6.4
uvicorn==0.29.0
